fix(yarn): spawn in shell on Windows
Must spawn `yarn` within a shell on Windows for the `yarn`
executable to be resolved correctly. Otherwise a ENOENT
error is received because the `yarn` command can't be found.

Spawning a comamnd in a shell on Linux should behave the same,
though there may be a small increase in time required to
spawn the process.