fix(gitlab): gracefully handle PR merge rejections (#2701)
Currently any GitLab PR merge rejection results in a Renovate ERROR being thrown. This includes 401 which is due to permissions:
![image](https://user-images.githubusercontent.com/6311784/47613176-25d72980-da8a-11e8-84cf-7441a50312ef.png)

Although it's unideal for someone to configure automerge but not grant Renovate adequate permissions to merge to master, it definitely should be gracefully handled and not throwing errors to logs.