fix(gitlab): close and reopen MR while rebasing to prevent errorneous GitLab CI message
Currently, using Renovate with GitLab makes GitLab CI to show the error message
"Could not connect to the CI server. Please check your settings and try again".
This issue is [a known bug of GitLab](https://gitlab.com/gitlab-org/gitlab-ce/issues/41545) and
the current workaround is to close and reopen the merge request.

This PR implements the aforementioned workaround.