feat: Add `rebaseConflictedPrs` configuration
`rebaseConflictedPrs` is used to enable or disable auto-rebase
in case of merge conflicts with the master branch. It is `true`
by default which means branches will be rebased if there are conflicts
in a PR. By setting it to `false`, Renovate no longer will rebase it
with the master branch if there are merge conflicts.

Refer #4184