fix: `ignorePaths` should not explicitly check for substring matches
BREAKING CHANGE: `ignorePaths` should not explicitly check for substring matches

Before this change, `ignorePaths` and `includePaths` are inconsistent:
* `ignorePaths` allows substring matches (or glob matches)
* `includePaths` requires exact string matches (or glob matches).

Having `ignorePaths` specifically check for substring matching is undesirable because:
* Substring matching can be achieved with glob matching if desired
* Substring matching can result in ignoring more paths than intended

I realize this is a breaking change that could result in renovatebot managing additional paths so I'm open to alternatives if that's too risky.

EXAMPLE:

I want to only ignore the top-level package.json in a repository.

Files:
* `package.json`
* `frontend/package.json`

Configuration:
* `"ignorePaths": ["package.json"]`

Unfortunately, due to substring matching, this matches all `package.json` files in the repository.

If I actually wanted to match all `package.json` files in the repository, I should have intentionally used the glob pattern `**/package.json`.

As a workaround, I can use `"includePaths": ["frontend/**"]`, but this is not as scalable / correct, as I would have to update it if I added more folders later.