fix: use package-lock as shrinkwrap when maintaining shrinkwrap
Fixes https://github.com/renovatebot/renovate/issues/4772

When running the `lockFileMaintenance`, the lock file is removed before it gets recreated. When the lock file is not present, npm creates a `package-lock.json` by default, so we need to rename it to `npm-shrinkwrap.json` when we want to get an updated version.