fix(gradle): qualifier falsely replaced [WIP]
In gradle qualifier dependencies can be used. e.g. testBuilder, when used in combination with contraints, the version can be ommited.  e.g: mysql:mysql-connector-java::mockServer is a valid definition. It's version is determined by a contraint

 dependencies {
            implementation "mysql:mysql-connector-java::mockServer"
            constraints {
                implementation "mysql:mysql-connector-java:6.0.5"
            }
 }

renovate updates this when a new version is found. mysql:mysql-connector-java::mockServer -> mysql:mysql-connector-java:7.0.0 which it shouldn't.

If added a testcase, but cannot do the actual fix.