refactor(docker-compose): add check that YAML is object
There is a test ('returns null for non-object YAML') for this behaviour,
but it was relying on the loop iterating over individual characters in
the resulting string and not finding anything useful in them,
which seems more like an accident than the way things were intentionally
structured. So, this adds an explicit check for this case.