refactor(docker-compose): remove seemingly unnecessary if statement
This was previously necessary to get the test 'returns null for
non-object YAML' to pass, but that seems to have been accidental and was
fixed in the previous commit.

This does change the behaviour of this function, but not (I think) in a way
which will cause issues: it will now return an empty array rather than null
in the case of a compose file that has no services with an image, but doing
so makes sense: the file was extracted successfully, it's just that there's
nothing of interest to be found in it.