fix(pip): fix compatibility with some simple pip indexes
Some "simple"-style pip indexes do not conform to the previous
assumption in this method, namely that they must end in `/+simple/` if
they provide only that style of API. Gemfury private indexes, for
example, are assigned as::

    pypi.fury.io/<username>

But offer only a simple API, eg. `pypi.fury.io/<username>/<pkg>/json`
will 404.

This changeset ensures Renovate works with indexes like this by adding a
fallback when the json API 404s to attempt the simple API. This is not
an optimal solution, but I could not find any way to be certain we can
determine whether the API will be json or simple in advance of querying
it.