feat(internal): use git to determine branch modification status
Uses git log to determine if the Renovate’s gitAuthor was the author of the most recent commit per branch. If it is, then the branch/PR is considered unmodified.

Closes #6693

BREAKING CHANGE: If Renovate shares a git “identity” with another real or bot user who may commit to Renovate’s branches, Renovate will no longer detect that as a modified branch. It is recommended to use an account or gitAuthor that is unique to Renovate, or ensure that the shared identity never commits to a Renovate branch.