changelog/github: Fix `files.length` condition
After using `shift()` in the line above, the array is one item shorter than before. So to detect if there are multiple candidates we need to check if there are any items left in the array at all. If we compare using `> 1` we check if there are 3+ candidates, by changing it to `!== 0` we can check if there are 2+ candidates.