feat: kustomize: support `newName` in `image`
kustomize allows overriding not just the version (`newTag`), but also the registry via `newName`. When a `newName` is present, k8s will look for the image in the location defined. If the overridden registry follows a different release cycle (e.g. has a delay for newly released images) - things will fail, i.e. `newTag` must be available for the `newName`, not the old `name`.