feat(azure): implemented azure branch status checks for automerge (#7398)
* feat(azure): implemented azure branch status checks for automerge

Implemented checks for branch statuses for azure, using the latest branch gitStatusState to determine if it's pending, failing or successful. This is required to properly automerge, as it was always pending prior.
Added tests for scenarios with various results and updated relevant existing tests.

fixes #7392

* feat(azure): finished implementing status checks and also setting of a branch status

Finished implementation according to how the other platforms work as much as possible.
Implemented the setting of a branch status, using the createCommitStatus api from azure.
Created util classes to handle the way azure translates named contexts with slashes to objects and reverse.
Tests for above changes.

Closes #7392

* test(azure): add test for uncovered lines to bring coverage back to 100

* chore(azure): fix the log message to be accurate and in the right spot and add a trace log to the end of the set status call

* Update lib/platform/azure/index.ts

Co-authored-by: Jamie Magee <JamieMagee@users.noreply.github.com>

* Update lib/platform/azure/index.ts

Co-authored-by: Jamie Magee <JamieMagee@users.noreply.github.com>
Co-authored-by: Rhys Arkins <rhys@arkins.net>