Improve handling of git submodules
This includes a couple of changes:

- Added a new boolean config option cloneSubmodules (Defaults to `true` to be a non-breaking change)
- Update submodules using `checkout` instead of `submdoule update --remote`
- Remove usage of simple-git `silent()` as it's deprecated

I wasn't able update a submodule using `update-index` and put the change in 'Changes not staged for commit'. I'm open to suggestions here. I have however changed the call from `git submodule update --init --remote` to a simple `git checkout`. The `--remote` flag updates the submodule to the latest commit at the remote, so there was a (very small chance) of a race condition here

Closes #7472