crate datasource: use memCache + privateCacheDir
This introduces a "privateCacheDir", which is simply
`${cacheDir}/__renovate-private-cache`, and is wiped between every
repository that's renovated, as per https://github.com/renovatebot/renovate/pull/8083#issuecomment-758622865

It's used by the crate datasource to clone the index of private
registries.

Additionally, within a single repo, the memCache is now used to store
"clone paths" for those private registries, to avoid re-cloning when
checking many dependencies from the same registry in the same
repository.

The memCache is also wiped between every repository being renovated,
this ensures no private information is leaked.