Only git-clone private registries if trustLevel=high
This changes `ReleasesConfigBase` to extend `RenovateConfig`, which was
already effectively the case everywhere except for the `go` datasource,
which is also fixed by this commit.

Git clones of untrusted repositories now only happen if `trustLevel` is
specifically set to `high`. It won't happen if it's set to `low` or, if
by chance, it's not set at all.