fix(poetry): fix support for private hosts
Adds a missing piece from #6994, which introduced support for private
package indexes to the poetry/artifacts.ts command. Unfortunately, this
only gives us support for places where we are running the literal
`poetry update` command in order to update the dependency, but does not
support the earlier extraction flow, eg. where we check the indexes for
new versions. By including the auth here, we ensure that packages are
searched for correctly in private indexes.

Behaviour before this changeset:
* if a package is present both in a private and public index, updates
  will only be found in the public index; but when doing the update, the
  private and public index will both be searched
* if a package is present in only a private index, it will never be
  updated

Behaviour with this changeset:
* private indexes are searched normally, so any updated package versions
  there will trigger a renovate operation to update that package