fix(manager/gitlabci): Update local include files
If local files are imported from the GitLab CI YAML file, the
dependencies will be recognized but not updated by renovate.
Renovate needs to respect the 'include:local' imports and update the
depedependencies in those files.

`extractPackageFile` is replaced by `extractAllPackageFiles` since the
functionality of those two functions are the same.

Related to #6745
Fixes #6713