fix: race condition when merging Azure PR
Azure can return 200 OK, but PR will still have status 1 (active) after updating to status 3 (closed).
That can cause issues with Azure not being able to finish merging the PR due to renovatebot going ahead and deleting the branch.
So add code to check for the expected status and wait+retry if needed.