Remove 20 page limit when fetching repo tags
On certain custom docker registries (`quay`..) each response will
contain 50 results and will require looping and fetching the next pages.
Currently the loop is limited to 20 pages, this means that only 1000
tags at most can be returned. This has the potential to break when a
repo has more than that.