Handle prereleases published with Poetry
Poetry accepts a union of SemVer and PEP440, normalizing certain
prerelease tags like alpha -> a, preview -> rc in its internal
representation.  Since packages published to PyPI use PEP440 versions,
this means that a version requirement "^0.8.0-alpha.1" should be
updated to "^0.8.0-alpha.2" under the `replace` rule when a version
"0.8.0a2" is published to PyPI.

To accomplish this, update the poetry versioning scheme to explicitly
parse versions using the same regex and normalization rules as
poetry itself, and reemit them in SemVer format before kicking down
to the npm versioning scheme. The normalization is applied in
reverse before returning.