fix(docker): handle redundant port in image
The docker data source was having trouble with images that have a
redundant port on them. Like putting `:443` on a docker image to get
this `artifactory.example.com:443/docker/alpine:3.13.1`

Having that redundant `:443` is pretty common for people who use
Artifactory specifically, since Artifactory puts it on everything in the
GUI.

Because of the way `URL` parses urls with a redundant port on it, the
string substitution wasn't working correctly since `url.origin` was
removing the port.

Fixes #11227