fix(cargo): Use proper cargo-update invocation
`cargo update --package somepkg` can fail if there are multiple versions
of that package in the project. There was an existing workaround, that
didn't seem to work for me.

`cargo update --workspace` is designed to be run after upgrading
`Cargo.toml` - it won't mess with transitive dependencies. We only
run `cargo update` (with no flags) when we're doing lockfile
maintenance, which _is_ supposed to upgrade transitive dependencies.

Closes https://github.com/renovatebot/renovate/issues/12332