feat(gradle): treat the preview qualifier as unstable (#12450)
This is an add-on to PR #12461 where we special cased the `-preview`
qualifier to be treated as unstable (and equivalent to `-rc`) in the
Maven versioning module.

Given that the Maven and Gradle versioning specs are very similar,
with one exception related to the ordering, I think it makes sense
to align them here.

Add `preview` to the list of unstable qualifiers for Gradle.