#13092 - Clarify digest upgrade ignore message
(discussion started in https://github.com/renovatebot/renovate/discussions/13092)

As @rarkins clarified, digest updates will resume if the PR title or branch name are updated. I have modified the comment so that this will be communicated to the user.

One question I have: Will this accurately describe behaviour in other cases?
I guess a docker image with the tag now pointing at a new image would be a `digest` upgrade? Would renovate according to the same rules in these cases?