fix: Do not drop bundler platforms on lock file maintenance
The bundler manager drops the Gemfile.lock before doing lock file
maintenance but bundler stores additional platforms for resolving only
in the lock file, such as `x86_64-darwin-20`. These additional platforms
and the resolved gems for these platforms are lost when the lock file
deleted. The newly generated platform will only contain the
`x86_64-linux` (or whatever the current platform is when renovate is
run).

This commit stops the manager from removing the lock file and instead
calls `bundler lock --update` to update the existing lock file
in-place.

Fixes #14156