Avoid replacing specific yarn constraint
When a yarn constraint has already been extracted by NPM, e.g. from
`packageManager` or `engine` in `package.json`, that version should not
be overridden by a generic range based on the lock file version.

If a specific yarn version is requested, that one should be used, even
if the lock file needs to change.