If the comment blocks merging, create thread instead of comment
Gitlab has the possibility to create either a standard comment, or a thread. A thread can be resolved, and, depending on the configuration of the repository, can even block the merge itself. By using a thread when the merge request indicates that it probably should not be merged as-is, the developer has to perform an addional action before the merge request can be merged. This makes it less likely that the developer missed the warning message.

@see https://docs.gitlab.com/ee/user/discussions/