feat: add renovate-schema.json to published package
tools like vscode have great support for includeing a `"$schema"`
property in your json for code completion and validation.

usually that `$schema` prop points to
`https://docs.renovatebot.com/renovate-schema.json` but this is
always the *latest* version of renovate.

in self-hosted environments not always the latest version is used,
so this patch adds the schema file to the bundled package for
easier reuse.

closes #13835