feat(tekton): support official bundle resolver
The functionality provided by the tektoncd/resolution project has been
added directly to the main tektoncd/pipelines project. Thus,
tektoncd/resolution is now deprecated.

As part of this move, one of the attributes in the CRD change names from
"resource" to "params". This commit drops support for the deprecated
"resource" attribute, and adds support for the officially supported
"params" attribute.

The docs have also been updated to reference the official documentation
instead of the dreprecated GitHub repository.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>