feat(manager/gomod): capture indirect dependencies for Go Modules
To allow folks to be able to discover the dependencies available through
indirect (transitive) dependencies for Go modules, we can parse them as
a separate `depType`.

This also makes sure that it's clear that `indirect` dependencies do not
get updated by Renovate.

This introduces a new `Package.enabled` field to allow us to -
longer-term - allow config-driven tuning for these.

Closes #19416.