feat: flag to control whether PRs can be edited by maintainers if `forkToken`is set
Platforms like GitHub expose options to control whether
a PR can be modified by maintainers.

See e.g. https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/allowing-changes-to-a-pull-request-branch-created-from-a-fork

This commit introduces a way to control this in Renovate when Renovate
operates in a fork. The option will default to `true` for backwards
compatibility.

Users may use this option to workaround issues as outlined in: #16657.