feat(github): respect new `prAllowMaintainerEdits` option when `forkToken` is set
This commit ensures the GitHub platform logic respects the
new `prallowMaintainerEdits` option when `forkToken` is set and
Renovate operates in a fork.

Whenever a PR is created with `forkToken` being set, depending on the
option value, the GitHub API post request will have the respective
option `maintainer_can_modify` option set.

This option was always enabled in fork mode before this commit,
and without fork mode enabled, the option is a noop because maintainers
have access to upstream branches regardless.

Users may want to leverage the option and disable it to avoid
issues with modifications in Renovate fork accounts, as described
in: #19771

Fixes #19771.