Add documentation for how to add a gitPrivateKey
Maybe because I'm not very familiar with GPG signing, but I really struggled with this step when I configured it on a gitlab runner.  I think the main issue I ran into was using the correct real name and email address, and possibly replacing all new lines with `\n`.  

I hope this makes it more clear for the next person.