index.ts: don't always pick the first release for OSV-based rules based on last_affected.
This is only relevant if a package rule is based on last_affected, e.g. "> 29.0", for regular fixed versions the list is already condensed to just one item at this point.

The version of the first entry here is compatible version-wise. It may hold a version suffix though which could be incompatible, e.g. 30.0-android, whereas 30.0-jre would be correct.