refactor: move `github-releases` datasource to `github-release-attachments`
The `github-releases` datasource currently computes a digest
based on existing digests that may point to release attachments.

This is not what the datasource should do for the majority of users
and for this reason the attachment "logic" is moved into a new
datasource (for backwards compatibility), with a new name of
`github-release-attachments`.