refactor: re-add `github-releases` datasource with improved `getDigest` logic
In the previous commit, the `github-releases` datasource has been
renamed to `github-release-attachments`. The `github-releases` datasource
`getDigest` method was rather magical and depended/required an existing
digest and version. This makes it difficult to use in some situations
where pinning is enabled but no initial digest is explicitly known. In
such cases, Renovate just silently stoped updating: see. #20160.

Additionally, some managers like `bazel` never pass a `digest` because
they handle the digest on their own. i.e. for Bazel updates with github
releases, the actual artifact checksum is used as digest and needs to
use an algorithm as determined by the manager.

The current approach of `getDigest` involved consulting the current
digest and iterating through the releases of the `currentValue` to
figure out which artifact/attachcment is used for the digest. The artifact
is then mapped back to the new release (with some trickery to also ignore
version names in the filename..). This is somewhat reasonable but the
data source generally does not have a clear insight into what GitHub
artifact(s) are used by the manager. Or, alternatively, a manager might
not even want the digest of an individual artifact. e.g. as seen in:
https://github.com/renovatebot/renovate/discussions/19550

To solve this, the existing datasource with this attachment logic is
was moved into `github-release-attachments`, while the more general
`github-releases` datasource is updated to use the underlying Git tag
for digest computation of a release entry.

BREAKING CHANGE: Regex Manager configurations relying on the
`github-release` data-source with digests will have different digest
semantics. The digest will always correspond to the underlying Git SHA
of the release/version now. The old behavior can be preserved by
switching to the `github-release-attachments` datasource.

Fixes #20160. Fixes #19552.