feat(template): add ifEquals helper
As an example of a use case for this is that I define package sources
using [purls](https://github.com/package-url/purl-spec) where the
`pkg:github` type defaults to use `datasource=github-releases`. I also
allow overriding the datasource via comments, such as `#
renovate=datasource=git-refs`. When doing so, I also have the need to
modify the `packageName` to be compatible with the `git-refs`
datasource, for example:

```yaml
id: pkg:github/renovatebot/renovate@116735f13
```

This will end up producing a renovate dependency with roughly the
following structure:

```json
{
  "datasource": "git-refs",
  "packageName": "renovatebot/renovate"
}
```

In my `regexManager`, I'd like to modify `packageName` to include
https://github.com/ to make it compatible with `git-refs`:

```json
"packageNameTemplate": "{{#ifEquals datasource 'git-refs'}}https://github.com/{{packageName}}{{else}}{{packageName}}{{/if}}",
```

For this, I'd need the ability to check if a value is equal to
something (`#ifEquals`), hence this PR.