fix: Test time difference with time zone
By testing without a time zone, the original test worked differently
based on the local time zone of the machine running the test. This is
presumably because of a summer time/winter time difference, because
changing the date from `2020-10-10` to `2020-05-10` without setting the
time zone also makes the test pass on my machine (Pacific/Auckland time
zone).