fix(datasource/helm): Digests are blank on Gitlab
In #23432 support was added for digests. This introduces a regression
when used in combination with the Helm repository on Gitlab.

In the `index.yaml` generated by Gitlabs Helm Repository the digest is left
blank. For example:

```
apiVersion: v2
entries:
  blank-digest:
  - apiVersion: v2
    created: '2023-09-05T13:25:19.046604000Z'
    description: A Helm chart with a blank digest
    digest:
    name: blank-digest
    urls:
    - https://charts.helm.sh/stable/packages/ambassador-2.7.0.tgz
    version: 2.7.0
```

As a result `newDigest: release.digest` will set the `newDigest`
to `null` and this causes the update to be removed in
`workers/repository/process/lookup/index.ts`.