feat(manager/bazel): recognize Bazel rules that begin with an underscore
When using Bazel, some users may wish to replace rules such as
`http_archive` with their own versions that behave slightly differently
or provide functionality that the regular versions do not.

When writing custom rules or functions, it is common to prefix their
names with an underscore to make them private. Therefore, this change
extends the Bazel manager to recognize rules beginning with an
underscore, e.g. `_http_archive` as well as `http_archive`.