Add support for non-default package repos to the pip-compile manager
This adds logic to pass specified non-default package registries to the
pip-compile subprocess when updating the lockfile.  It needs these to
allow it to resolve indirect dependencies.

The registry URLs to pass to the external subprocess are provided by
adding a `registryUrls` array in the `pip-compile` configuration block
and credentials for private registries are found by matching against the
`hostRules`.

```json
{
  "pip-compile": {
    ...
    "registryUrls": ["https://pypi.my.domain/pypi/simple"],
    ...
  },
  "hostRules": [
    {
      "matchHost": "pypi.my.domain",
      "username": ...,
      "password": ...
    }
  ]
}
```