refactor(manager/pip-compile) Use python keychain module for credentials
This refactors the pip-compile manager to use the python keychain module
along with the keychains.envvars module to pass credentials to pip.
Doing it this way means there's a significantly lower likelyhood of
accidentally leaking credentials into the output file, since credentials
looked up via the keychain module won't ever be written to --index-url
or --extra-index-url directives in the output file.

Because of this lower likelyhood of leaking credentials, I relaxed the
usage of --no-emit-index-url to only automatically apply when
credentials are found in index URLs in environment variables.