chore(logging): log the new and current digest in the branch info
Also log the new and current digest in the debug branch info log
to be able to know exactly what the branches are updating.

This is especially useful for branches that are just updating digests.

My use case for this is being able to parse the renovate log output
to extract what branches and PRs are exactly updating.