Update correct lock files when updating a dependency in multiple inputs
This applies the fix from #27319 to non-lockFileMaintanence updates so
that the correct lockFiles attribute gets passed to the manager's
updateArtifacts function when updating single dependencies.  Without
this change, the same lockFiles attribute gets passed along with every
packageFileName, which causes the pip-compile manager to update the same
lock file multiple times when a dependency is updated in multiple input
files.