feat(manager/pip-compile): Support -r dependencies between files
Pip-compile supports including other requirements files via the -r flag.
The pip_requirements manager already extracts those flags, but this adds
logic to the pip-compile manager to make sure dependent files' lock
files get updated when a package is updated in a file that is included
via a -r directive.