Add tests for paginated PRs
This adds two tests for paginated PRs. The first test is for paginated PRs where
gitlab returns the expected link and the second test is for paginated PRs where
gitlab returns a different link.
The first test could have been added bevore this PR and isn't changed by this PR.