refactor(workers/repository): Reverse the update logic in getUpdatedPackageFiles
This reverses the logic in getUpdatedPackageFiles so that it runs all of
the updates for each manager sequentially.

That is, instead of
```
for each package file
    for each manager
        manager.update(package file)
```
it will do
```
for each manager with updated package files
    for each package file
        manager.update(package file)
```

This refactor is the basis for a fix to #28518.  To resolve that issue,
the package files need to be re-sorted to be in the same order that the
manager extracted them.  Since a package file can be relevant to
multiple managers, the list of package files needs to be re-sorted for
each manager.  That isn't possible with the existing logic.