feat(pip_requirements) Extract flags from package files with no deps
The pip_requirements manager is used by the pip-compile manager for
extracting info from requirements.in files.  This change makes
pip_requirements return any flags it extracted from its input file even
if there are no dependencies in the file.  This allows a user to provide
a requirements.in file with nothing but `--extra-index-url` flags as a
second input file when using pip-compile to compile dependencies in a
setup.py file so that Renovate can look those index URLs up in its
`hostRules` and pass appropriate credentials to pip-compile.