fix(pip-compile): Correctly report errors when a lock file is unchanged
The pip-compile manager can update multiple lock files when a single
input file changes, however the manager would short-circuit and return
null as soon as a single lock file remained unchanged after being
re-compiled.

This was causing two issues:
1. Updates to a later lock file could be missed.  This is fairly
   unlikely since the manager trys not to re-compile lock files for
   which the input haven't changed.
2. Errors in one lock file update could be swolloed if a later lock file
   update succeeded but made no changes.  This scenario was quite
   likely, especially in the case where one lock file depends on another
   since the first failure will result in an unchanged upstream lock
   file.

This removes the short-circuit behavior so that the manager will
re-compile all of the lock files that depend on the changed input file
no matter what.