feat(gomod): add support for package replacements
To make it possible for folks using Go modules to be able to take
advantage of Renovate's package replacements, we should add support to
the `gomod` manager.

This adds support for managing the following `updateType`s:

- minor/patch
- major
- digest

For `major` and `digest` updates, we need to support the case that we're
performing a replacement, so add an additional check for the
`updateType=replacement` and whether there's a `newMajor` or
`newDigest`.

Because we're now managing multiple `depName`s, we should rename this to
`fromPackageName` and `toPackageName`, following how we've done this
with the `hermit` manager.

Additionally, as we're replacing the dependency's name, we want to have
a separate regex capture group for `depName`, which we can also use for
determining whether there are quotes around the dependency.

As part of #24883.