chore(regex): add additional logging for `isValidDependency` failures
When writing regexes to match custom file formats, it may not be obvious
why a valid regex may not be resulting in a match.

This is often because we have written a perfectly correct regex which
has matched the line(s), but then our `isValidDependency` check fails.

To make it easier to discover this, we can add a TRACE log, and note
that we will discard the dependency, giving suitable context for what is
being rejected.