fix(local): don't validate `reconfigure` branch
As noted in #31007, validating the `renovate/reconfigure` branch when
running with the local platform ends up in an error reported to the
user:

```
TypeError: Cannot read properties of undefined (reading 'renovate/reconfigure')
```

This error is because we're trying to do something we shouldn't do with
the local platform - interact with branches - and as there's no way for
the user to fix it, this error isn't useful.

Users aren't impacted by anything as part of this behaviour, nor can
they do anything to resolve this, so this error reporting isn't useful.

To avoid this, we can perform an early return (and log) to skip this
error condition.