Escape special characters in bundler usernames
In [this discussion](https://github.com/renovatebot/renovate/discussions/32185#discussioncomment-11101283)
we demonstrated a problem with how Renovate passes usernames to the
external bundler command. If the username contains special characters,
such as the @ sign in an email address, bundler would fail with the
error text

```
Please CGI escape your usernames and passwords before setting them for
authentication.
```

Renovate does its own escaping elsewhere, such as when querying versions
prior to opening PRs. So it isn't possible to update the username at the
config level. Instead, this PR places the responsibility of properly
escaping the username on the bundler/host-rules module.