feat(github-action): retain spaces in extracted values
This update ensures that Renovate retains the
number of whitespaces already present in the
GitHub Action YAML file.

If there are no existing comments indicating the
version number (e.g., when initially pinning a
dependency), Renovate's behaviour remains
unchanged (that is, it will insert a single space
between the commit hash and the `#`).

Signed-off-by: JP-Ellis <josh@jpellis.me>