feat: Add support for Bun's text-format lockfile
Bun 1.2 will introduce and switch by default to a text-format verison of
the lockfile. Where currently they use a binary version named
`bun.lockb`, this will become `bun.lock` and be plaintext:

https://bun.sh/blog/bun-lock-text-lockfile

This can be switched on by default in-advance since Bun 1.1.40 and so
will appear in the wild as of now:

https://bun.sh/blog/bun-v1.1.40#new-install-savetextlockfile-in-bunfig-toml

This commit adds support for both lockfile formats by supporting by
file-suffixes.